Missing archiver when deleting selected contacts
When deleting one or more contacts via radio button selection the contacts do get deleted but the HTML table doesn't update because the Archiver isn't being passed back to "index.html" in the `render_template()` function. This causes an error in "index.html" for the line `{% if archiver.status() == "Waiting" %}` that can be seen via the Flask debug server. If you manually reload the page the contacts disappear but I don't think that is how you wanted to example to behave.

[HYPERMEDIA SYSTEMS: The Server Side for Delete Selected Contacts](https://hypermedia.systems/book/more-htmx-patterns/#_the_server_side_for_delete_selected_contacts)

The book is great! Thanks for sharing it!